Rewrite in Go to avoid runtime deps
Depending on busybox-static means that any service that depends on
chef/mlsa will have busybox in their path. This PR replaces the shell
script with a small Go program that should be equivalent.

One bummer is that the go binary is actually pretty large even after
stripping. One alternative would be to just use C, but this seemed
more easily maintainable by a broader group of Chef developers over
time.

Signed-off-by: Steven Danna <steve@chef.io>